// packages/plugin-pos-api/src/models/definitions/configs.ts
export interface IEbarimtConfig {
  companyName: string;
  ebarimtUrl: string;
  checkCompanyUrl: string;
  hasVat: boolean;
  hasCitytax: boolean;
  districtCode: string;
  companyRD: string;
  defaultGSCode: string;
  vatPercent: number;
  cityTaxPercent: number;
  footerText: string;
}

export interface IQPayConfig {
  url: string;
  callbackUrl: string;
  username: string;
  password: string;
  invoiceCode: string;
}

interface IConfigColors {
  [key: string]: string;
}

interface IUIOptions {
  colors: IConfigColors;
  logo: string;
  bgImage: string;
  favIcon: string;
  receiptIcon: string;
  texts: IConfigColors;
}

interface ISyncInfo {
  id: string;
  date: Date;
}

interface ICatProd {
  _id: string;
  categoryId: string;
  productId: string;
}

export interface IConfig {
  name: string;
  description?: string;
  brandId: string;
  tagIds?: string[];
  productDetails: string[];
  adminIds: string[];
  cashierIds: string[];
  beginNumber: string;
  maxSkipNumber: number;
  kitchenScreen: any;
  waitingScreen: any;
  kioskMachine?: any;
  formSectionTitle?: string;
  formIntegrationIds?: string[];
  token?: string;
  uiOptions: IUIOptions;
  ebarimtConfig?: IEbarimtConfig;
  qpayConfig?: IQPayConfig;
  syncInfo?: ISyncInfo;
  catProdMappings: ICatProd[];
  initialCategoryIds: string[];
  kioskExcludeProductIds: string[];
}

export interface IConfigDocument extends Document, IConfig {
  _id: string;
}